poker v0.1.0 Poker

A poker game module implementing hand checking, tie breaking, and outputting.

Functions

baseHand(hand)

baseHand(List.t()) :: List.t()

Takes the remainder when divided by 13 for each element of a list.

Parameters

- hand: List representing a poker hand.

deal(intList)

Deals cards based on input and sends winning hand to output function.

Parameters

- intList: List of 10 unique integers from 1-52 representing two poker hands.

find_n_OfKind(list, num)

getPairList(hand)

getPairList(List.t()) :: List.t()

Extracts the duplicate elements of a list.

Parameters

- hand: List representing a poker hand.

getType(hand)

getType(List.t()) :: Integer.t()

Uses hand checking functions to return the highest type.

Parameters

- hand: List representing a poker hand.

is_flush(hand)

is_flush(List.t()) :: Boolean.t()

Checks if a hand is a flush.

Parameters

- hand: List representing a poker hand.

is_fourOfKind(hand)

is_fourOfKind(List.t()) :: Boolean.t()

Checks if a hand is a four of a kind.

Parameters

- hand: List representing a poker hand.

is_fullHouse(hand)

is_fullHouse(List.t()) :: Boolean.t()

Checks if a hand is a full house.

Parameters

- hand: List representing a poker hand.

is_pair(hand)

is_pair(List.t()) :: Boolean.t()

Checks if a hand is a pair.

Parameters

- hand: List representing a poker hand.

is_royalFlush(hand)

is_royalFlush(List.t()) :: Boolean.t()

Checks if a hand is a royal flush.

Parameters

- hand: List representing a poker hand.

is_straight(hand)

is_straight(List.t()) :: Boolean.t()

Checks if a hand is a straight.

Parameters

- hand: List representing a poker hand.

is_straightFlush(hand)

is_straightFlush(List.t()) :: Boolean.t()

Checks if a hand is a striaght flush.

Parameters

- hand: List representing a poker hand.

is_threeOfKind(hand)

is_threeOfKind(List.t()) :: Boolean.t()

Checks if a hand is a three of a kind.

Parameters

- hand: List representing a poker hand.

is_twoPair(hand)

is_twoPair(List.t()) :: Boolean.t()

Checks if a hand is a two pair.

Parameters

- hand: List representing a poker hand.

normalizeHand(hand)

normalizeHand(List.t()) :: List.t()

Adds 13 to each element of 1 or 0 in a list helping to evaluate highcards.

Parameters

- hand: List representing a poker hand.

output(hand)

output(List.t()) :: List.t()

output(List.t()) :: List.t()

Returns the properly formatted winning poker hand.

Parameters

- hand: List representing the winning poker hand.

sortDesc(hand)

sortDesc(List.t()) :: List.t()

Sorts a list in descending order helping to evaluate highcards and suits.

Parameters

- hand: List representing a poker hand.

tieBreak(hand1, hand2, type)

tieBreak(List.t(), List.t(), Integer.t()) :: List.t()

Uses tie breaking functions to return winning hand

Parameters

- hand1: List representing player one's poker hand.
- hand2: List representing player two's poker hand.
- type: Integer representing the type of hand in common.

tie_fourOfKind(hand1, hand2)

tie_fourOfKind(List.t(), List.t()) :: List.t()

Tie brakes two four of a kind type hands.

Parameters

- hand1: List representing player one's poker hand.
- hand2: List representing player two's poker hand.

tie_highcard(hand1, hand2)

tie_highcard(List.t(), List.t()) :: List.t()

Finds the hand with the highest card value or suit.

Parameters

- hand1: List representing player one's poker hand.
- hand2: List representing player two's poker hand.

tie_pair(hand1, hand2)

tie_pair(List.t(), List.t()) :: List.t()

Tie brakes two one pair type hands.

Parameters

- hand1: List representing player one's poker hand.
- hand2: List representing player two's poker hand.

tie_straight(hand1, hand2)

tie_straight(List.t(), List.t()) :: List.t()

Tie brakes two straight type hands.

Parameters

- hand1: List representing player one's poker hand.
- hand2: List representing player two's poker hand.

tie_threeOfKind(hand1, hand2)

tie_threeOfKind(List.t(), List.t()) :: List.t()

Tie brakes two three of a kind type hands.

Parameters

- hand1: List representing player one's poker hand.
- hand2: List representing player two's poker hand.